+++ backend/db/sqlc/user.sql.go
@@ -65,7 +65,7 @@ SELECT id, username, hashed_password, first_name, last_name, email, phone_number
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetUser(ctx context.Context, id int32) (User, error) {
+func (q *Queries) GetUser(ctx context.Context, id int64) (User, error) {
 	row := q.db.QueryRowContext(ctx, getUser, id)
 	var i User
 	err := row.Scan(
